Made final changes to replace API with front end
