+++ index.js
@@ -1,6 +1,6 @@
 const app = require('./app');
 
-const port = 3000;
+const port = 3001;
 
 app.listen(port, () => {
   console.log(`Listening on port ${port}`);
