+++ package.json
@@ -7,7 +7,8 @@
     "start": "nodemon index.js",
     "test": "jest --forceExit",
     "build-data": "psql -d loungeo -f schemas/schema.sql && psql -d loungeo -f schemas/data_uploader.sql && psql -d loungeo -f schemas/data_modifier.sql",
-    "see-data": "psql -d loungeo -f schemas/test.sql"
+    "see-data": "psql -d loungeo -f schemas/test.sql",
+    "k6": "k6 run spec/script.js"
   },
   "repository": {
     "type": "git",
