+++ schemas/schema.sql
@@ -11,7 +11,7 @@ DROP TABLE IF EXISTS char_name_vote_join CASCADE;
 
 
 CREATE TABLE photos (
-	photos_id integer GENERATED BY DEFAULT AS IDENTITY UNIQUE PRIMARY KEY,
+	photos_id integer GENERATED BY DEFAULT AS IDENTITY,
 	url varchar(255) DEFAULT null,
 	review_id integer NOT NULL
 ) WITH (
